requirements.txt
boto3
pyyaml
attrs
cmdln
